requirements.txt
folium
requests
fastapi
pydantic[dotenv]
uvicorn
python-dotenv